1、调试支付宝的当面付 2、完善order模块和goods模块 3、新增订单超时之后自动恢复商品的状态功能

+++ irh-service/irh-service-goods/src/main/java/top/imuster/goods/service/impl/ProductRecommendService.java
@@ -17,7 +17,7 @@ public class ProductRecommendService {
     @Autowired
     ProductRecommendDao productRecommendDao;
 
-    public ProductRecommendDto findOne(Integer userId){
+    public ProductRecommendDto findOne(Long userId){
         return productRecommendDao.findByUserId(userId);
     }
